* lang_t.slang_of * is now lang_t*. * lang_[n]lookup return lang_t*. As for now, it is converted to ordinal   number with varr_index, but later this may be converted to lang_t.id   as it will allow full control of langs varr (e.g. deletion). For now,   remove land_destroy as it will lead to confusion in mlstr's. * rfile_t.len is off_t, while mmap/mmunmap calls take only size_t args.   cast them to size_t (which happened before quietly), but a better   solution may be to make them size_t and check for overflow in rfile_open * [Q] why do hashstr() and hashcasestr() declared in old style in str.c?   Or were they just copied from somewhere? * strkey_search_cb smashed, it weren't used and was duplicated by   vstr_search_cb. * fix a potential coredump in varr_cpy, when v_data->e_cpy == NULL   (we have such varr's, but luckily seem not to copy them :) * [Q] can we speed up vo_foreach() a bit, if we untag unprocessed objects   starting with (rv), not (vo) ? AFAIU all objects before (rv) are already   untagged. * common #defines and declarations in cc_lex.l/cc_parse.y moved to cc_lex.h * comm.c really does not need global d_next. Moreover, this could cause   skipping next descriptor when the current is closed (since it was   incremented twice, in game_loop_unix() and close_descriptor()). * mp triggers should be reworked, to be queued and executed after all other   processing is done (e.g. from game_loop_unix or a separate update handler).   In the way they are implemented now there are chances to do the Wrong   Thing. E.g. in comm_act.c mp_act_trigger is called and then   write_to_buffer() to one of CHAR_DATA subject to that buffer. During   execution of mp_act_trigger char can be extracted, can't it? * Introduce _DC() macro to temporary shut up warnings where the   implementation should be reworked * main module now compiles with -DEXTRA_CHECKS, make them default for   building bin/sog. * -Werror disabled, as yacc and lex both generate files uncompatible with   BDECFLAGS. 2 PRs were filed after this cleanup, 1 (bin/28362)is now   waiting for MFC of skeleton.c in yacc, 1 (bin/28364) is being negotiated   with vendor of lex. * more const'ification of functions in handler.c may be still needed, but   it compiles for now cleanly and that will go in separate commit * [Q] Does anyone have something against changing 'You hit the reality...'   to 'You MANGLE the reality. Reality truth [*] POWER HITS [*] you!'   (Seems that we inherited this message from Diku or something alike, but   they have different damage nouns)
git-svn-id: file:///home/fjoe/old-repos/svn/husky/trunk/sog@2882 1738b366-ca78-dd11-ac90-000e8e03bdb1

+++ src/include/str.h
@@ -23,7 +23,7 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  *
- * $Id: str.h,v 1.16 2001-06-23 17:17:14 fjoe Exp $
+ * $Id: str.h,v 1.17 2001-06-24 10:50:40 avn Exp $
  */
 
 #ifndef _STR_H_
@@ -63,8 +63,8 @@ bool	str_prefix	(const char *astr, const char *bstr);
 bool	str_infix	(const char *astr, const char *bstr);
 bool	str_suffix	(const char *astr, const char *bstr);
 
-int hashstr(const char *s, int maxn, int hashs);
-int hashcasestr(const char *s, int maxn, int hashs);
+size_t hashstr(const char *s, size_t maxn, size_t hashs);
+size_t hashcasestr(const char *s, size_t maxn, size_t hashs);
 
 /**
  * Compare two pointers to strings (case-insensitive)
